refactor: 음식점 리스트 컴포넌트 화
1. App에서 음식점 리스트 템플릿, 랜더 담당 함수 분리
2. Restaurant.js, RestaurantList.js  컴포넌트파일 생성
